+++ lib/searchkick/query.rb
@@ -12,14 +12,14 @@ module Searchkick
       :took, :error, :model_name, :entry_name, :total_count, :total_entries,
       :current_page, :per_page, :limit_value, :padding, :total_pages, :num_pages,
       :offset_value, :offset, :previous_page, :prev_page, :next_page, :first_page?, :last_page?,
-      :out_of_range?, :hits, :response, :to_a, :first
+      :out_of_range?, :hits, :response, :to_a, :first, :scroll
 
     def initialize(klass, term = "*", **options)
-      unknown_keywords = options.keys - [:aggs, :body, :body_options, :boost,
+      unknown_keywords = options.keys - [:aggs, :block, :body, :body_options, :boost,
         :boost_by, :boost_by_distance, :boost_by_recency, :boost_where, :conversions, :conversions_term, :debug, :emoji, :exclude, :execute, :explain,
         :fields, :highlight, :includes, :index_name, :indices_boost, :limit, :load,
-        :match, :misspellings, :model_includes, :offset, :operator, :order, :padding, :page, :per_page, :profile,
-        :request_params, :routing, :scope_results, :select, :similar, :smart_aggs, :suggest, :total_entries, :track, :type, :where]
+        :match, :misspellings, :models, :model_includes, :offset, :operator, :order, :padding, :page, :per_page, :profile,
+        :request_params, :routing, :scope_results, :scroll, :select, :similar, :smart_aggs, :suggest, :total_entries, :track, :type, :where]
       raise ArgumentError, "unknown keywords: #{unknown_keywords.join(", ")}" if unknown_keywords.any?
 
       term = term.to_s
@@ -39,6 +39,7 @@ module Searchkick
       @misspellings = false
       @misspellings_below = nil
       @highlighted_fields = nil
+      @index_mapping = nil
 
       prepare
     end
@@ -56,9 +57,19 @@ module Searchkick
     end
 
     def params
+      if options[:models]
+        @index_mapping = {}
+        Array(options[:models]).each do |model|
+          # there can be multiple models per index name due to inheritance - see #1259
+          (@index_mapping[model.searchkick_index.name] ||= []) << model
+        end
+      end
+
       index =
         if options[:index_name]
           Array(options[:index_name]).map { |v| v.respond_to?(:searchkick_index) ? v.searchkick_index.name : v }.join(",")
+        elsif options[:models]
+          @index_mapping.keys.join(",")
         elsif searchkick_index
           searchkick_index.name
         else
@@ -71,6 +82,7 @@ module Searchkick
       }
       params[:type] = @type if @type
       params[:routing] = @routing if @routing
+      params[:scroll] = @scroll if @scroll
       params.merge!(options[:request_params]) if options[:request_params]
       params
     end
@@ -98,7 +110,9 @@ module Searchkick
       # no easy way to tell which host the client will use
       host = Searchkick.client.transport.hosts.first
       credentials = host[:user] || host[:password] ? "#{host[:user]}:#{host[:password]}@" : nil
-      "curl #{host[:protocol]}://#{credentials}#{host[:host]}:#{host[:port]}/#{CGI.escape(index)}#{type ? "/#{type.map { |t| CGI.escape(t) }.join(',')}" : ''}/_search?pretty -H 'Content-Type: application/json' -d '#{query[:body].to_json}'"
+      params = ["pretty"]
+      params << "scroll=#{options[:scroll]}" if options[:scroll]
+      "curl #{host[:protocol]}://#{credentials}#{host[:host]}:#{host[:port]}/#{CGI.escape(index)}#{type ? "/#{type.map { |t| CGI.escape(t) }.join(',')}" : ''}/_search?#{params.join('&')} -H 'Content-Type: application/json' -d '#{query[:body].to_json}'"
     end
 
     def handle_response(response)
@@ -116,8 +130,10 @@ module Searchkick
         misspellings: @misspellings,
         term: term,
         scope_results: options[:scope_results],
-        index_name: options[:index_name],
-        total_entries: options[:total_entries]
+        total_entries: options[:total_entries],
+        index_mapping: @index_mapping,
+        suggest: options[:suggest],
+        scroll: options[:scroll]
       }
 
       if options[:debug]
@@ -166,7 +182,7 @@ module Searchkick
     end
 
     def retry_misspellings?(response)
-      @misspellings_below && response["hits"]["total"] < @misspellings_below
+      @misspellings_below && Searchkick::Results.new(searchkick_klass, response).total_count < @misspellings_below
     end
 
     private
@@ -218,6 +234,7 @@ module Searchkick
       per_page = (options[:limit] || options[:per_page] || 10_000).to_i
       padding = [options[:padding].to_i, 0].max
       offset = options[:offset] || (page - 1) * per_page + padding
+      scroll = options[:scroll]
 
       # model and eager loading
       load = options[:load].nil? ? true : options[:load]
@@ -377,7 +394,7 @@ module Searchkick
               queries_to_add.concat(q2)
             end
 
-            queries.concat(queries_to_add)
+            queries << queries_to_add
 
             if options[:exclude]
               must_not.concat(set_exclude(exclude_field, exclude_analyzer))
@@ -392,9 +409,10 @@ module Searchkick
 
             should = []
           else
+            # higher score for matching more fields
             payload = {
-              dis_max: {
-                queries: queries
+              bool: {
+                should: queries.map { |qs| {dis_max: {queries: qs}} }
               }
             }
 
@@ -410,6 +428,22 @@ module Searchkick
         where = (options[:where] || {}).dup
         if searchkick_options[:inheritance] && (options[:type] || (klass != searchkick_klass && searchkick_index))
           where[:type] = [options[:type] || klass].flatten.map { |v| searchkick_index.klass_document_type(v, true) }
+        end
+
+        models = Array(options[:models])
+        if models.any? { |m| m != m.searchkick_klass }
+          warn "[searchkick] WARNING: Passing child models to models option throws off hits and pagination - use type option instead"
+
+          # uncomment once aliases are supported with _index
+          # see https://github.com/elastic/elasticsearch/issues/23306
+          # index_type_or =
+          #   models.map do |m|
+          #     v = {_index: m.searchkick_index.name}
+          #     v[:type] = m.searchkick_index.klass_document_type(m, true) if m != m.searchkick_klass
+          #     v
+          #   end
+
+          # where[:or] = Array(where[:or]) + [index_type_or]
         end
 
         # start everything as efficient filters
@@ -469,7 +503,7 @@ module Searchkick
       pagination_options = options[:page] || options[:limit] || options[:per_page] || options[:offset] || options[:padding]
       if !options[:body] || pagination_options
         payload[:size] = per_page
-        payload[:from] = offset
+        payload[:from] = offset if offset > 0
       end
 
       # type
@@ -483,11 +517,21 @@ module Searchkick
       # merge more body options
       payload = payload.deep_merge(options[:body_options]) if options[:body_options]
 
+      # run block
+      options[:block].call(payload) if options[:block]
+
+      # scroll optimization when interating over all docs
+      # https://www.elastic.co/guide/en/elasticsearch/reference/current/search-request-scroll.html
+      if options[:scroll] && payload[:query] == {match_all: {}}
+        payload[:sort] ||= ["_doc"]
+      end
+
       @body = payload
       @page = page
       @per_page = per_page
       @padding = padding
       @load = load
+      @scroll = scroll
     end
 
     def set_fields
@@ -660,20 +704,9 @@ module Searchkick
     def set_boost_by_indices(payload)
       return unless options[:indices_boost]
 
-      if below52?
-        indices_boost = options[:indices_boost].each_with_object({}) do |(key, boost), memo|
-          index = key.respond_to?(:searchkick_index) ? key.searchkick_index.name : key
-          # try to use index explicitly instead of alias: https://github.com/elasticsearch/elasticsearch/issues/4756
-          index_by_alias = Searchkick.client.indices.get_alias(index: index).keys.first
-          memo[index_by_alias || index] = boost
-        end
-      else
-        # array format supports alias resolution
-        # https://github.com/elastic/elasticsearch/pull/21393
-        indices_boost = options[:indices_boost].map do |key, boost|
-          index = key.respond_to?(:searchkick_index) ? key.searchkick_index.name : key
-          {index => boost}
-        end
+      indices_boost = options[:indices_boost].map do |key, boost|
+        index = key.respond_to?(:searchkick_index) ? key.searchkick_index.name : key
+        {index => boost}
       end
 
       payload[:indices_boost] = indices_boost
@@ -710,7 +743,7 @@ module Searchkick
     def set_highlights(payload, fields)
       payload[:highlight] = {
         fields: Hash[fields.map { |f| [f, {}] }],
-        fragment_size: below60? ? 30000 : 0
+        fragment_size: 0
       }
 
       if options[:highlight].is_a?(Hash)
@@ -821,7 +854,7 @@ module Searchkick
     # TODO id transformation for arrays
     def set_order(payload)
       order = options[:order].is_a?(Enumerable) ? options[:order] : {options[:order] => :asc}
-      id_field = below60? ? :_uid : :_id
+      id_field = :_id
       payload[:sort] = order.is_a?(Array) ? order : Hash[order.map { |k, v| [k.to_s == "id" ? id_field : k, v] }]
     end
 
@@ -843,7 +876,12 @@ module Searchkick
         else
           # expand ranges
           if value.is_a?(Range)
-            value = {gte: value.first, (value.exclude_end? ? :lt : :lte) => value.last}
+            # infinite? added in Ruby 2.4
+            if value.end.nil? || (value.end.respond_to?(:infinite?) && value.end.infinite?)
+              value = {gte: value.first}
+            else
+              value = {gte: value.first, (value.exclude_end? ? :lt : :lte) => value.last}
+            end
           end
 
           value = {in: value} if value.is_a?(Array)
@@ -1013,16 +1051,12 @@ module Searchkick
       k.sub(/\.(analyzed|word_start|word_middle|word_end|text_start|text_middle|text_end|exact)\z/, "")
     end
 
-    def below52?
-      Searchkick.server_below?("5.2.0")
-    end
-
-    def below60?
-      Searchkick.server_below?("6.0.0")
-    end
-
     def below61?
       Searchkick.server_below?("6.1.0")
+    end
+
+    def below70?
+      Searchkick.server_below?("7.0.0")
     end
   end
 end